+++ src/test/java/blackjack/domain/card/CardPickerTest.java
@@ -1,7 +1,5 @@
-package blackjack.domain;
+package blackjack.domain.card;
 
-import blackjack.domain.card.Card;
-import blackjack.domain.card.CardPicker;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
