+++ src/test/java/blackjack/domain/game/RefereeTest.java
@@ -1,9 +1,7 @@
-package blackjack.domain;
+package blackjack.domain.game;
 
 import blackjack.domain.card.Card;
 import blackjack.domain.card.CardPicker;
-import blackjack.domain.game.GameResult;
-import blackjack.domain.game.Referee;
 import blackjack.domain.gamer.Dealer;
 import blackjack.domain.gamer.Player;
 import org.junit.jupiter.api.BeforeEach;
